api/requirements.txt
alembic==1.13.3
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-Migrate==4.0.7
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
pkgutil-resolve-name==1.3.10
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.20.0
SQLAlchemy==2.0.36
typing-extensions==4.12.2
werkzeug==3.0.4
zipp==3.20.2
